In StartingUserInterface's CreateCommand method update the condition to check for a length of 3 instead of 2. Update the message sent by the CreateCommand when it doesn't have the correct number of parts to "Command not valid, Create requires a name and type of gradebook.". Set the value of gradeBook based on the value of the third substring from parts.
parts will be set by user input, the second substring is the name of the gradebook, the third substring is the type of gradebook, and later the fourth will be if grading is weighted or not.

Remove the instantiation from where we declare the gradeBook variable.
If the third substring's value is "standard" set gradeBook to a new instantiation of StandardGradeBook.
If the third substring's value is "ranked" set gradeBook to a new instantiation of RankedGradeBook.
If the third substring's value doesn't match either of the above write the value the user entered followed by " is not a supported type of gradebook, please try again" then escape the method.